+++ hybrid_core/hybrid_core/CHANGELOG.md
@@ -9,3 +9,9 @@
 ## 0.0.3
 
 * `macOS` Fix the issue that `OS#atLeastVersion` always returns true.
+
+## 0.1.0
+
+* Move `AndroidOS` to `Android`.
+* Move `DarwinOS` to `Darwin`.
+* Move `DarwinOSVersion` to `DarwinVersion`.