+++ hybrid_core/hybrid_core_android/CHANGELOG.md
@@ -5,3 +5,7 @@
 ## 0.0.2
 
 * Fix `OSPlatform.instance` throws UnimplementedError.
+
+## 0.1.0
+
+* Move `AndroidOSImpl` to `AndroidImpl`.